move sidebar items to sidebar component
